+++ src/components/resizable_container/helpers.test.ts
@@ -45,8 +45,10 @@ describe('sizesOnly', () => {
         getSizePx: () => 10,
         minSize: ['0px', '0px'],
         isCollapsed: false,
+        initSize: panel.size,
         prevSize: 0,
         position: 'middle',
+        canGrow: false,
         ...panel,
       };
       return out;